added folders for game info types refactored GameLevelsLinks component
