+++ src/components/shared/gameLevelsLinks/GameLevelsLinks.tsx
@@ -1,26 +1,34 @@
 'use client';
 
 import LevelsLinksContainer from './LevelsLinksContainer';
-import React, { useEffect, useState } from 'react';
-import { MemoTheMeloGame, PitchCatchGame } from '@/types';
+import { useEffect, useState } from 'react';
+import { MemoTheMeloGame, PitchCatchGame, LevelStatus } from '@/types';
 import StageTitle from './StageTitle';
 import StageLevelsContainer from './StageLevelsContainer';
 import LevelLink from './LevelLink';
 import { useSession } from 'next-auth/react';
-import { LevelStatus } from '@/types';
 import { Locale } from '@/i18n.config';
+import type { Session } from 'next-auth';
 
 type Props = {
     levelsData: MemoTheMeloGame | PitchCatchGame;
     lang: Locale;
     game: string;
 };
+
 type UserProgressEntry = {
     [key: string]: LevelStatus;
 };
 
+type ExtendedSession = Session & {
+    user: Session['user'] & {
+        id: string; // Change the type of id according to your setup
+    };
+};
+
+
 const GameLevelsLinks = ({ levelsData, lang, game }: Props) => {
-    const { data: session } = useSession();
+    const { data: session } = useSession() as { data: ExtendedSession | null };
     const [userProgress, setUserProgress] = useState<UserProgressEntry>({});
 
     const getStatus = (
@@ -45,12 +53,9 @@ const GameLevelsLinks = ({ levelsData, lang, game }: Props) => {
 
     useEffect(() => {
         // Fetch user progress from the backend using the user's ID (session.id)
-        console.log('game', game);
         const fetchUserProgress = async () => {
             try {
-                //@ts-ignore
                 const response = await fetch(
-                    //@ts-ignore
                     `/api/games/${game}/${session?.user?.id}`,
                 );
                 const data = await response.json();
