+++ src/types/games/index.ts
@@ -1,0 +1,4 @@
+
+export * from './memoTheMelo';
+export * from './memoBlocks';
+export * from './pitchCatch';
