+++ src/types/games/memoTheMelo/index.ts
@@ -1,0 +1,8 @@
+export type MemoTheMeloLevel = {
+  pitchOptions: string[];
+  melody: number[];
+};
+
+export type MemoTheMeloStage = MemoTheMeloLevel[];
+
+export type MemoTheMeloGame = MemoTheMeloStage[];
