+++ src/types/games/pitchCatch/index.ts
@@ -1,0 +1,10 @@
+export type PitchCatchLevel = PitchCatchQuestion[];
+
+export type PitchCatchStage = PitchCatchLevel[];
+
+export type PitchCatchGame = PitchCatchStage[];
+
+export type PitchCatchQuestion = {
+  currPitch: string[];
+  userOptions: string[][];
+};
